+++ shared/src/main/java/com/gabstra/myworkoutassistant/shared/sets/WeightSet.kt
@@ -2,4 +2,4 @@ package com.gabstra.myworkoutassistant.shared.sets
 
 import java.util.UUID
 
-data class WeightSet(override val id: UUID, val reps: Int, val weight: Float) : Set(id)
+data class WeightSet(override val id: UUID, val reps: Int, val weight: Double) : Set(id)
